+++ changelog.md
@@ -1,6 +1,24 @@
 # Changelog
 **All notable changes to this project will be recorded here**
 
+# Version 2.0
+**(Oct 17 2019)**
+
+# Changes
+* Better handling of HTTPS
+* Automatic URL and port detection -- if url uses http or https
+* Write errors to separate file
+* Better error reporting -- More descriptive errors
+* Option for nocolors
+* Suppress errors when running
+* Only print certain status codes (can request multiple status codes at the same time.)
+
+# Future Changes
+* Ability to set timeout for connect and read, rather than one set timeout for both
+* Print descriptive definition of status codes
+* Handle URL invalid schemes better
+* Possibly the ability to send X number of requests at a time, wait N number of seconds, send X number again, etc.
+
 # Version 1.0
 
 # Notes